# be/src/docs/swagger.yaml
openapi: 3.0.0
info:
  title: Car Insurance API
  description: |
    Complete API documentation for Car Insurance Management System.
    
    This API provides endpoints for:
    - User authentication with OTP verification
    - Insurance policy management
    - Payment processing with Stripe
    - Notification management
    
    ## Authentication
    
    The API uses JWT (JSON Web Tokens) for authentication. After successful OTP verification,
    clients receive a JWT token that must be included in the Authorization header for protected endpoints.
    
    Format: `Authorization: Bearer <jwt_token>`
    
    ## Error Handling
    
    All error responses follow a consistent format:
    ```json
    {
      "message": "Error description",
      "code": "ERROR_CODE",
      "timestamp": "2026-02-17T10:30:00Z"
    }
    ```
    
  version: 1.0.0
  contact:
    name: Car Insurance Team
    email: team@carinsurance.com
  license:
    name: MIT
servers:
  - url: https://letmbe.ashydune-d638a33c.westus2.azurecontainerapps.io
    description: Development server
  - url: https://api.yourdomain.com
    description: Production server
...
paths:
...
  /api/notifications:
    get:
      tags:
        - Notifications
      summary: Get user notifications
      description: Retrieve all notifications for the authenticated user
      operationId: getNotifications
      security:
        - bearerAuth: []
      parameters:
        - name: limit
          in: query
          description: Maximum number of notifications to return
          schema:
            type: integer
            default: 20
            minimum: 1
            maximum: 100
          example: 10
        - name: skip
          in: query
          description: Number of notifications to skip (for pagination)
          schema:
            type: integer
            default: 0
            minimum: 0
          example: 0
        - name: read
          in: query
          description: Filter by read status
          schema:
            type: boolean
          example: false
      responses:
        '200':
          description: List of notifications
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                    example: true
                  data:
                    type: array
                    items:
                      $ref: '#/components/schemas/Notification'
                  total:
                    type: integer
                    example: 5
                  limit:
                    type: integer
                    example: 20
                  skip:
                    type: integer
                    example: 0
        '401':
          description: Unauthorized - Invalid or missing token
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
              examples:
                noToken:
                  value:
                    message: Authorization token is missing
                    code: MISSING_TOKEN
                invalidToken:
                  value:
                    message: Invalid or expired token
                    code: INVALID_TOKEN
        '500':
          description: Server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
...
components:
  schemas:
...
    Notification:
      type: object
      properties:
        _id:
          type: string
          description: Notification ID
          example: 507f1f77bcf86cd799439011
        userId:
          type: string
          description: Related user ID
          example: 507f1f77bcf86cd799439012
        type:
          type: string
          description: Notification type
          enum: [policy_update, payment_reminder, payment_confirmation, policy_expiry, quote_available]
          example: payment_reminder
        title:
          type: string
          example: Payment Due
        message:
          type: string
          example: Your monthly insurance premium is due on February 20th
        read:
          type: boolean
          description: Whether notification has been read
          example: false
        actionUrl:
          type: string
          description: URL to take action on this notification
          format: uri
          example: /payments/policy-123
        metadata:
          type: object
          description: Additional notification data
          example:
            policyId: 507f1f77bcf86cd799439011
            amount: 1200
        createdAt:
          type: string
          format: date-time
          example: '2026-02-17T10:30:00Z'
        updatedAt:
          type: string
          format: date-time
          example: '2026-02-17T10:30:00Z'
    Error:
      type: object
      required:
        - message
        - code
      properties:
        message:
          type: string
          description: Error message
          example: An error occurred
        code:
          type: string
          description: Error code for client-side handling
          example: ERROR_CODE
        details:
          type: object
          description: Additional error details
          example:
            field: email
            value: invalid
        timestamp:
          type: string
          format: date-time
          example: '2026-02-17T10:30:00Z'
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: |
        JWT token obtained from the /api/auth/verify-otp endpoint.
        
        Example:
        ```
        Authorization: Bearer eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9...
        ```
...